Moved searchSaga and searchSlice to common/Navigation/SearchBar, updated imports in store and rootSaga
