+++ src/common/Navigation/SearchBar/searchSaga.js
@@ -1,25 +1,25 @@
 import { all, call, delay, put, select, takeLatest } from "redux-saga/effects";
 import { pageNumberFromURL, selectInputValue, selectPage, selectPath, setData, setInputValue, setPath, setTotalPages, setTotalResult } from "./searchSlice";
+import { getSearch } from "../../../API/getSearch";
+
 
 function* fetchDataHandler() {
     try {
         yield delay(120);
-        const [inputValue, path, page] = yield all([
+        const [query, path, page] = yield all([
             select(selectInputValue),
             select(selectPath),
             select(selectPage),
         ]);
-       
         if (query !== null) {
             const searchResults =
-            yield call(getSearch, inputValue, path, page);
+            yield call(getSearch, query, path, page);
 
             yield put(setData(searchResults));
             yield put(setTotalResult(searchResults));
             yield put(setTotalPages(searchResults));
 
         }
-
     } catch (error) {
     };
 };
@@ -27,4 +27,3 @@ function* fetchDataHandler() {
 export function* searchSaga() {
     yield takeLatest([setInputValue, pageNumberFromURL], fetchDataHandler);
 };
-
