+++ src/core/rootSaga.js
@@ -2,7 +2,7 @@ import { all } from "redux-saga/effects";
 import { movieListSaga } from "../features/movies/MovieList/movieListSaga";
 import { peopleListSaga } from "../features/people/PeopleList/peopleListSaga";
 import { movieSaga } from "../features/movies/MovieDetails/movieSaga";
-import { searchSaga } from "../features/search/searchSaga";
+import { searchSaga } from "../common/Navigation/SearchBar/searchSaga";
 
 
 export default function* rootSaga() {
