+++ src/core/store.js
@@ -4,7 +4,7 @@ import rootSaga from "./rootSaga";
 import movieListReducer from "../features/movies/MovieList/movieListSlice";
 import peopleListReducer from "../features/people/PeopleList/peopleListSlice";
 import movieReducer from "../features/movies/MovieDetails/movieSlice";
-import searchReducer from "../features/search/searchSlice";
+import searchReducer from "../common/Navigation/SearchBar/searchSlice";
 
 const sagaMiddleware = createSagaMiddleware();
